refactor: can pass string to the left label title
