+++ packages/ui-components/src/Inputs/InputTitle/InputTitle.tsx
@@ -1,19 +1,9 @@
-import { above } from '@webb-dapp/ui-components/utils/responsive-utils';
 import React from 'react';
-import styled, { css } from 'styled-components';
+import styled from 'styled-components';
 
 import { FlexBox } from '../..';
-
-export const sharedLabelCss = css`
-  font-family: 'Work Sans', sans-serif;
-  font-size: 14px;
-  line-height: 20px;
-  letter-spacing: 0.02em;
-
-  ${above.xs`
-    font-size: 16px;
-  `}
-`;
+import { sharedLabelCss } from '../shared';
+import { TextLabel } from '../TextLabel/TextLabel';
 
 const InputTitleWrapper = styled(FlexBox).attrs({
   width: '100%',
@@ -25,14 +15,16 @@ const InputTitleWrapper = styled(FlexBox).attrs({
 `;
 
 export type InputTitleProps = {
-  leftLabel?: React.ReactNode;
+  leftLabel?: string | React.ReactNode;
   rightLabel?: React.ReactNode;
 };
 
 export const InputTitle: React.FC<InputTitleProps> = ({ leftLabel, rightLabel }) => {
   return (
     <InputTitleWrapper>
-      <div style={{ marginRight: 'auto' }}>{leftLabel}</div>
+      <div style={{ marginRight: 'auto' }}>
+        {typeof leftLabel === 'string' ? <TextLabel value={leftLabel} /> : leftLabel}
+      </div>
       <div style={{ marginLeft: 'auto' }}>{rightLabel}</div>
     </InputTitleWrapper>
   );
